exposure-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
kafka-python==2.0.2
boto3==1.34.0